Make evalexception's context size configurable from the outside.
By default, it's fixed to 2 lines of context, which can be not enough in some
cases.
This is now configurable per-deployment, and default to 2 lines, as before.

Also, don't depend on paste.deploy just for parameters convertion. The
functions are simple enough to inline them into the source, and to avoid the
dependency just for this.